+++ app/Services/Ftp.php
@@ -426,4 +426,11 @@ class Ftp extends Directory_Listing_Base implements Service {
 		// return whether this file type is allowed (false) or not (true).
 		return ! in_array( $mime_type['type'], Helper::get_allowed_mime_types(), true );
 	}
+
+	/**
+	 * Initialize WP CLI for this service.
+	 *
+	 * @return void
+	 */
+	public function cli(): void {}
 }